+++ example/lib/main.dart
@@ -5,6 +5,7 @@ import 'package:example/code_block.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_confetti/flutter_confetti.dart';
 import 'package:syntax_highlight/syntax_highlight.dart';
+import 'package:google_fonts/google_fonts.dart';
 
 // late final Highlighter _dartLightHighlighter;
 late final Highlighter _dartDarkHighlighter;
@@ -43,6 +44,7 @@ class MainApp extends StatefulWidget {
 
 class _MainAppState extends State<MainApp> {
   final controller = ConfettiController();
+  List<ConfettiController> killableControllerList = [];
 
   @override
   Widget build(BuildContext context) {
@@ -64,15 +66,10 @@ class _MainAppState extends State<MainApp> {
                   highlighter: _dartDarkHighlighter,
                   onTap: () {
                     ///BEGIN
-
                     Confetti.launch(
                       context,
                       options: const ConfettiOptions(
-                        particleCount: 100,
-                        spread: 70,
-                        y: 0.6,
-                        // flat: true,
-                      ),
+                          particleCount: 100, spread: 70, y: 0.6),
                     );
 
                     ///END
@@ -89,8 +86,11 @@ class _MainAppState extends State<MainApp> {
 
                     Confetti.launch(
                       context,
-                      options:
-                          ConfettiOptions(angle: randomInRange(55, 125), spread: randomInRange(50, 70), particleCount: randomInRange(50, 100).toInt(), y: 0.6),
+                      options: ConfettiOptions(
+                          angle: randomInRange(55, 125),
+                          spread: randomInRange(50, 70),
+                          particleCount: randomInRange(50, 100).toInt(),
+                          y: 0.6),
                     );
 
                     ///END
@@ -121,12 +121,22 @@ class _MainAppState extends State<MainApp> {
                       Confetti.launch(
                         context,
                         options: ConfettiOptions(
-                            particleCount: count, startVelocity: 30, spread: 360, ticks: 60, x: randomInRange(0.1, 0.3), y: Random().nextDouble() - 0.2),
+                            particleCount: count,
+                            startVelocity: 30,
+                            spread: 360,
+                            ticks: 60,
+                            x: randomInRange(0.1, 0.3),
+                            y: Random().nextDouble() - 0.2),
                       );
                       Confetti.launch(
                         context,
                         options: ConfettiOptions(
-                            particleCount: count, startVelocity: 30, spread: 360, ticks: 60, x: randomInRange(0.7, 0.9), y: Random().nextDouble() - 0.2),
+                            particleCount: count,
+                            startVelocity: 30,
+                            spread: 360,
+                            ticks: 60,
+                            x: randomInRange(0.7, 0.9),
+                            y: Random().nextDouble() - 0.2),
                       );
                     });
 
@@ -145,20 +155,71 @@ class _MainAppState extends State<MainApp> {
                         gravity: 0,
                         decay: 0.94,
                         startVelocity: 30,
-                        colors: [Color(0xffFFE400), Color(0xffFFBD00), Color(0xffE89400), Color(0xffFFCA6C), Color(0xffFDFFB8)]);
+                        colors: [
+                          Color(0xffFFE400),
+                          Color(0xffFFBD00),
+                          Color(0xffE89400),
+                          Color(0xffFFCA6C),
+                          Color(0xffFDFFB8)
+                        ]);
 
                     shoot() {
-                      Confetti.launch(context, options: options.copyWith(particleCount: 40, scalar: 1.2), particleBuilder: (index) => Star());
+                      Confetti.launch(context,
+                          options:
+                              options.copyWith(particleCount: 40, scalar: 1.2),
+                          particleBuilder: (index) => Star());
                       Confetti.launch(context,
                           options: options.copyWith(
                             particleCount: 10,
                             scalar: 0.75,
-                          ));
+                          ),
+                          particleBuilder: (index) => Star());
                     }
 
                     Timer(Duration.zero, shoot);
                     Timer(const Duration(milliseconds: 100), shoot);
                     Timer(const Duration(milliseconds: 200), shoot);
+
+                    ///END
+                  },
+                ),
+                CodeBlock(
+                  buttonText: 'Emoji',
+                  tip:
+                      'The demo uses google web fonts so the first call may take a few seconds to download the font.',
+                  highlighter: _dartDarkHighlighter,
+                  onTap: () {
+                    ///BEGIN
+
+                    const options = ConfettiOptions(
+                      spread: 360,
+                      ticks: 50,
+                      gravity: 0,
+                      decay: 0.94,
+                      startVelocity: 30,
+                    );
+
+                    shoot() {
+                      Confetti.launch(context,
+                          options: options.copyWith(
+                            particleCount: 40,
+                          ),
+                          particleBuilder: (index) => Emoji(
+                              emoji: '🍄',
+                              textStyle: GoogleFonts.notoColorEmoji()));
+                      Confetti.launch(context,
+                          options: options.copyWith(
+                            particleCount: 10,
+                          ),
+                          particleBuilder: (index) => Emoji(
+                                emoji: '️⚽',
+                                textStyle: GoogleFonts.notoColorEmoji(),
+                              ));
+                    }
+
+                    Timer(Duration.zero, shoot);
+                    Timer(const Duration(milliseconds: 200), shoot);
+                    Timer(const Duration(milliseconds: 400), shoot);
 
                     ///END
                   },
@@ -193,7 +254,12 @@ class _MainAppState extends State<MainApp> {
                       if (controller1 == null) {
                         controller1 = Confetti.launch(
                           context,
-                          options: const ConfettiOptions(particleCount: 2, angle: 60, spread: 55, x: 0, colors: colors),
+                          options: const ConfettiOptions(
+                              particleCount: 2,
+                              angle: 60,
+                              spread: 55,
+                              x: 0,
+                              colors: colors),
                           onFinished: (overlayEntry) {
                             if (isDone) {
                               overlayEntry.remove();
@@ -207,7 +273,12 @@ class _MainAppState extends State<MainApp> {
                       if (controller2 == null) {
                         controller2 = Confetti.launch(
                           context,
-                          options: const ConfettiOptions(particleCount: 2, angle: 120, spread: 55, x: 1, colors: colors),
+                          options: const ConfettiOptions(
+                              particleCount: 2,
+                              angle: 120,
+                              spread: 55,
+                              x: 1,
+                              colors: colors),
                           onFinished: (overlayEntry) {
                             if (isDone) {
                               overlayEntry.remove();
@@ -221,6 +292,34 @@ class _MainAppState extends State<MainApp> {
 
                     ///END
                   },
+                ),
+                CodeBlock(
+                  buttonText: 'Launch',
+                  tip: 'Launch then click the kill button.',
+                  highlighter: _dartDarkHighlighter,
+                  onTap: () {
+                    ///BEGIN
+
+                    final controller = Confetti.launch(
+                      context,
+                      options: const ConfettiOptions(
+                          particleCount: 100, spread: 70, y: 0.6),
+                    );
+
+                    /// call the kill method to kill the confetti
+                    /// controller.kill();
+
+                    ///END
+
+                    killableControllerList.add(controller);
+                  },
+                  otherButton: OutlinedButton(
+                      onPressed: () {
+                        for (var controller in killableControllerList) {
+                          controller.kill();
+                        }
+                      },
+                      child: const Text('Kill')),
                 ),
                 CodeBlock(
                   buttonText: 'Not Full Screen',
@@ -242,20 +341,22 @@ class _MainAppState extends State<MainApp> {
 
                     ///END
                   },
+                  overWidget: Positioned(
+                      right: 24,
+                      left: 24,
+                      bottom: 24,
+                      top: 82,
+                      child: Container(
+                        // height: 200,
+                        child: ClipPath(
+                          child: Confetti(
+                            controller: controller,
+                            options: const ConfettiOptions(
+                                particleCount: 100, spread: 70, y: 1),
+                          ),
+                        ),
+                      )),
                 ),
-                Container(
-                  decoration: BoxDecoration(
-                    color: Colors.black,
-                    borderRadius: const BorderRadius.all(Radius.circular(8)),
-                    border: Border.all(width: 1, color: Theme.of(context).primaryColor),
-                  ),
-                  child: ClipPath(
-                    child: Confetti(
-                      controller: controller,
-                      options: const ConfettiOptions(particleCount: 100, spread: 70, y: 1),
-                    ),
-                  ),
-                )
               ],
             ),
           );