+++ lib/src/confetti.dart
@@ -7,6 +7,7 @@ import 'package:flutter_confetti/src/confetti_physics.dart';
 import 'package:flutter_confetti/src/shapes/square.dart';
 import 'package:flutter_confetti/src/utils/glue.dart';
 import 'package:flutter_confetti/src/utils/launcher.dart';
+import 'package:flutter_confetti/src/utils/launcher_config.dart';
 import 'package:flutter_confetti/src/utils/painter.dart';
 import 'package:flutter_confetti/src/confetti_particle.dart';
 import 'package:flutter_confetti/src/shapes/circle.dart';
@@ -157,6 +158,12 @@ class _ConfettiState extends State<Confetti> with SingleTickerProviderStateMixin
     playAnimation();
   }
 
+  kill() {
+    for (var glue in glueList) {
+      glue.physics.kill();
+    }
+  }
+
   @override
   void initState() {
     super.initState();
@@ -171,7 +178,8 @@ class _ConfettiState extends State<Confetti> with SingleTickerProviderStateMixin
       );
     }
 
-    Launcher.load(widget.controller, launch);
+    Launcher.load(
+        widget.controller, LauncherConfig(onLaunch: launch, onKill: kill));
   }
 
   @override
@@ -180,7 +188,8 @@ class _ConfettiState extends State<Confetti> with SingleTickerProviderStateMixin
 
     if (widget.controller != oldWidget.controller) {
       Launcher.unload(oldWidget.controller);
-      Launcher.load(widget.controller, launch);
+      Launcher.load(
+          widget.controller, LauncherConfig(onLaunch: launch, onKill: kill));
     }
   }