+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+## 0.3.4
+
+- Add the `controller.kill()` method to kill the showing confetti
+
+## 0.3.3
+
+- Improve the render quality of the emoji
+
+## 0.3.2
+
+- Add the emoji shape
+
+## 0.3.1
+
+- Add the triangle shape by [@Imad Eddine](https://github.com/DidoHZ)
+
 ## 0.3.0
 
 - the `controller` argument of `Confetti` is required now